modify before deployment

+++ teamproject/backend/src/apis/column/column.resolver.ts
@@ -1,4 +1,3 @@
-import { UseGuards } from '@nestjs/common';
 import { Args, Mutation, Query, Resolver } from '@nestjs/graphql';
 import { GqlAuthAccessGuard } from 'src/common/auth/gql-auth.guard';
 import { CurrentUser, ICurrentUser } from 'src/common/auth/gql-user.param';
@@ -8,9 +7,22 @@ import { CreateColumnInput } from './dto/createColumn.input';
 import { UpdateColumnInput } from './dto/updateColumn.input';
 import { CoachColumn } from './entities/column.entity';
 
+import { Cache } from 'cache-manager';
+import { CACHE_MANAGER, Inject, UseGuards } from '@nestjs/common';
+import { Roles } from 'src/common/auth/gql-role.param';
+import { Role } from '../user/entities/user.entity';
+import { RolesGuard } from 'src/common/auth/gql-role.guard';
+
+// import { ElasticsearchService } from '@nestjs/elasticsearch';
+
 @Resolver()
 export class CoachColumnResolver {
-  constructor(private readonly coachColumnService: CoachColumnService) {}
+  constructor(
+    private readonly coachColumnService: CoachColumnService,
+
+    @Inject(CACHE_MANAGER)
+    private readonly cacheManager: Cache,
+  ) {}
   // useGuard
   // roleGuard
   @Query(() => [CoachColumn])
@@ -32,6 +44,11 @@ export class CoachColumnResolver {
       pageNum,
       itemcount,
     });
+  }
+
+  @Query(() => [CoachColumn])
+  async fetchSearchedColumnList(@Args('search') search: string) {
+    return await this.coachColumnService.findAllSearchArgsColumn({ search });
   }
 
   @Query(() => [CoachColumn])
@@ -61,15 +78,22 @@ export class CoachColumnResolver {
   }
 
   //---------login user only-------------
-  @UseGuards(GqlAuthAccessGuard)
+  @Roles(Role.COACH)
+  @UseGuards(GqlAuthAccessGuard, RolesGuard)
   @Query(() => [CoachColumn])
   async fetchMyColumn(@CurrentUser() currentUser: ICurrentUser) {
     return await this.coachColumnService.findMyColumn({ currentUser });
   }
 
-  async uploadColumnFile() {}
-
+  // async uploadColumnFile() {}
   @UseGuards(GqlAuthAccessGuard)
+  @Query(() => [ColumnLike])
+  async fetchColumnListWhatILike(@CurrentUser() currentUser: ICurrentUser) {
+    return await this.coachColumnService.findColumnListILike({ currentUser });
+  }
+
+  @Roles(Role.COACH)
+  @UseGuards(GqlAuthAccessGuard, RolesGuard)
   @Mutation(() => CoachColumn)
   async createColumn(
     @CurrentUser() currentUser: ICurrentUser,
@@ -81,7 +105,8 @@ export class CoachColumnResolver {
     });
   }
 
-  @UseGuards(GqlAuthAccessGuard)
+  @Roles(Role.COACH)
+  @UseGuards(GqlAuthAccessGuard, RolesGuard)
   @Mutation(() => CoachColumn)
   async updateColumn(
     @Args('columnId') columnId: string,
@@ -93,9 +118,13 @@ export class CoachColumnResolver {
     });
   }
 
-  @UseGuards(GqlAuthAccessGuard)
+  @Roles(Role.COACH)
+  @UseGuards(GqlAuthAccessGuard, RolesGuard)
   @Mutation(() => Boolean)
-  async deleteColumn(@Args('columnId') columnId: string) {
-    return await this.coachColumnService.delete({ columnId });
+  async deleteColumn(
+    @CurrentUser() currentUser: ICurrentUser,
+    @Args('columnId') columnId: string,
+  ) {
+    return await this.coachColumnService.delete({ columnId, currentUser });
   }
 }
